tests/api/requirements.txt
pytest==8.0.0
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0